Use instrinsicContentSize rather than manipulating autolayout constraints
The approach of handling autolayout by searching for an manipulating height constraints is error-prone. Additionally the fact that the search results table is outside the bounds of the KSTokenView doesn't work well when the KSTokenView is in a UIStackView.

This patch handles autolayout by giving the KSTokenView the correct, dynamic instrinsicContentSize and including the search results table within the bounds of the KSTokenView. This gives correct behaviour in all circumstances.

Existing projects where the KSTokenView is in Interface Builder should remove any height constraints on the KSTokenView and instead give it a placeholder intrinsic content size in IB with width "none" and height 30 (for example).

Also, KSTokenView now works just fine when clipsToBounds is true, which can give a small performance gain.